refactor category board
