+++ src/components/CategoryBoard/Leaderboard/ChapterRankCard.tsx
