+++ src/components/CategoryBoard/Leaderboard/ChapterRankMinicard.tsx
